requirements.txt
charset-normalizer==3.4.3
numpy==2.3.3
packaging==25.0
pdf2image==1.17.0
pillow==11.3.0
PyMuPDF==1.26.4
PyPDF2==3.0.1
pytesseract==0.3.13
reportlab==4.4.4
tqdm==4.67.1
